+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/common/buildkite_trigger_build_action.rb
@@ -28,7 +28,7 @@ module Fastlane
         )
 
         if response.state == 'scheduled'
-          UI.message("Successfully scheduled new build. You can see it at '#{response.web_url}'")
+          UI.success("Successfully scheduled new build. You can see it at '#{response.web_url}'")
         else
           UI.crash!("Failed to start job\nError: [#{response}]")
         end