+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/ios/ios_betabuild_prechecks.rb
@@ -63,12 +63,12 @@ module Fastlane
           FastlaneCore::ConfigItem.new(key: :base_version,
                                        env_name: 'FL_IOS_BETABUILD_PRECHECKS_BASE_VERSION',
                                        description: 'The version to work on', # a short description of this parameter
-                                       is_string: true,
-                                       optional: true), # true: verifies the input is a string, false: every kind of value),
+                                       type: String,
+                                       optional: true),
           FastlaneCore::ConfigItem.new(key: :skip_confirm,
                                        env_name: 'FL_IOS_BETABUILD_PRECHECKS_SKIPCONFIRM',
                                        description: 'Skips confirmation',
-                                       is_string: false, # true: verifies the input is a string, false: every kind of value
+                                       type: Boolean,
                                        default_value: false), # the default value if the user didn't provide one
           FastlaneCore::ConfigItem.new(key: :default_branch,
                                        env_name: 'FL_RELEASE_TOOLKIT_DEFAULT_BRANCH',