+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/ios/ios_update_metadata_source.rb
@@ -18,7 +18,6 @@ module Fastlane
         repo_clean = repo_status.empty?
         unless repo_clean
           Action.sh('git commit -m "Update metadata strings"')
-          Action.sh('git push')
         end
       end
 
@@ -42,7 +41,7 @@ module Fastlane
           FastlaneCore::ConfigItem.new(key: :po_file_path,
                                        env_name: 'FL_IOS_UPDATE_METADATA_SOURCE_PO_FILE_PATH',
                                        description: 'The path of the .po file to update',
-                                       is_string: true,
+                                       type: String,
                                        verify_block: proc do |value|
                                                        UI.user_error!("No .po file path for UpdateMetadataSourceAction given, pass using `po_file_path: 'file path'`") unless value && (!value.empty?)
                                                        UI.user_error!("Couldn't find file at path '#{value}'") unless File.exist?(value)
@@ -56,7 +55,7 @@ module Fastlane
           FastlaneCore::ConfigItem.new(key: :source_files,
                                        env_name: 'FL_IOS_UPDATE_METADATA_SOURCE_SOURCE_FILES',
                                        description: 'The hash with the path to the source files and the key to use to include their content',
-                                       is_string: false,
+                                       type: Hash,
                                        verify_block: proc do |value|
                                                        UI.user_error!("No source file hash for UpdateMetadataSourceAction given, pass using `source_files: 'source file hash'`") unless value && (!value.empty?)
                                                      end),