+++ lib/fastlane/plugin/wpmreleasetoolkit/helper/git_helper.rb
@@ -73,16 +73,15 @@ module Fastlane
         Action.sh('git', 'submodule', 'update', '--init', '--recursive')
       end
 
-      # Create a new branch named `branch_name`, cutting it from branch/commit/tag `from`, and push it
+      # Create a new branch named `branch_name`, cutting it from branch/commit/tag `from`
       #
       # If the branch with that name already exists, it will instead switch to it and pull new commits.
       #
       # @param [String] branch_name The full name of the new branch to create, e.g "release/1.2"
       # @param [String?] from The branch or tag from which to cut the branch from.
       #        If `nil`, will cut the new branch from the current commit. Otherwise, will checkout that commit/branch/tag before cutting the branch.
-      # @param [Bool] push If true, will also push the branch to `origin`, tracking the upstream branch with the local one.
       #
-      def self.create_branch(branch_name, from: nil, push: true)
+      def self.create_branch(branch_name, from: nil)
         if branch_exists?(branch_name)
           UI.message("Branch #{branch_name} already exists. Skipping creation.")
           Action.sh('git', 'checkout', branch_name)
@@ -90,22 +89,19 @@ module Fastlane
         else
           Action.sh('git', 'checkout', from) unless from.nil?
           Action.sh('git', 'checkout', '-b', branch_name)
-          Action.sh('git', 'push', '-u', 'origin', branch_name) if push
         end
       end
 
       # `git add` the specified files (if any provided) then commit them using the provided message.
-      # Optionally, push the commit to the remote too.
       #
       # @param [String] message The commit message to use
       # @param [String|Array<String>] files A file or array of files to git-add before creating the commit.
       #        Use `nil` or `[]` if you already added the files in a separate step and don't wan't this method to add any new file before commit.
       #        Also accepts the special symbol `:all` to add all the files (`git commit -a -m …`).
-      # @param [Bool] push If true, will `git push` to `origin` after the commit has been created. Defaults to `false`.
       #
-      # @return [Bool] True if commit and push were successful, false if there was an issue during commit & push (most likely being "nothing to commit").
+      # @return [Bool] True if commit was successful, false if there was an issue (most likely being "nothing to commit").
       #
-      def self.commit(message:, files: nil, push: false)
+      def self.commit(message:, files: nil)
         files = [files] if files.is_a?(String)
         args = []
         if files == :all
@@ -115,7 +111,6 @@ module Fastlane
         end
         begin
           Action.sh('git', 'commit', *args, '-m', message)
-          Action.sh('git', 'push', 'origin', 'HEAD') if push
           return true
         rescue
           return false
@@ -179,6 +174,23 @@ module Fastlane
       #
       def self.fetch_all_tags
         Action.sh('git', 'fetch', '--tags')
+      end
+
+      # Returns the current git branch, or "HEAD" if it's not checked out to any branch
+      # Can NOT be replaced using the environment variables such as `GIT_BRANCH` or `BUILDKITE_BRANCH`
+      #
+      # `fastlane` already has a helper action for this called `git_branch`, however it's modified
+      # by CI environment variables. We need to check which branch we are actually on and not the
+      # initial branch a CI build is started from, so we are using the `git_branch_name_using_HEAD`
+      # helper instead.
+      #
+      # See https://docs.fastlane.tools/actions/git_branch/#git_branch
+      #
+      # @return [String] The current git branch, or "HEAD" if it's not checked out to any branch
+      #
+      def self.current_git_branch
+        # We can't use `other_action.git_branch`, because it is modified by environment variables in Buildkite.
+        Fastlane::Actions.git_branch_name_using_HEAD
       end
 
       # Checks if a branch exists locally.